src/requirements.txt
Flask==2.2.3
Pillow==9.5.0
transformers==4.28.1
pytesseract==0.3.10
Requests==2.28.2
# torch==2.0.0          # Install these manually (i.e., if you need CUDA or not)
# torchvision==0.15.1
flask_swagger_ui==4.11.1
timm==0.6.13
